+++ src/main/java/ru/practicum/shareit/booking/BookingController.java
@@ -1,6 +1,8 @@
 package ru.practicum.shareit.booking;
 
+import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.shareit.booking.dto.BookingDto;
@@ -20,99 +22,49 @@ import java.util.List;
 @RestController
 @RequestMapping(path = "/bookings")
 @Validated
+@AllArgsConstructor
 public class BookingController {
     private final BookingService bookingService;
     private final UserService userService;
     private final ItemService itemService;
 
-    @Autowired
-    public BookingController(BookingService bookingService, UserService userService, ItemService itemService) {
-        this.bookingService = bookingService;
-        this.userService = userService;
-        this.itemService = itemService;
-    }
-
     @PostMapping
-    public BookingDto createBooking(@RequestBody @Valid BookingDtoForCreate bookingDtoCreate,
+    @ResponseStatus(HttpStatus.CREATED)
+    @Validated
+    public BookingDto createBooking(@RequestBody @Valid @NotNull BookingDtoForCreate bookingDtoCreate,
                                     @RequestHeader(value = "X-Sharer-User-Id") @NotNull Long userId) {
-        if (bookingDtoCreate == null) {
-            throw new BadParametrException("при бронировании не было передано тело запроса");
-        }
-        if (bookingDtoCreate.getStart().isBefore(LocalDateTime.now()) ||
-                bookingDtoCreate.getEnd().isBefore(bookingDtoCreate.getStart()) ||
-                bookingDtoCreate.getEnd().equals(bookingDtoCreate.getStart())) {
-            throw new BadParametrException("Неверный формат даты старта и окончания бронирования: Start Date: "
-                    + bookingDtoCreate.getStart() + " End Date: " + bookingDtoCreate.getEnd() + " Now: "
-                    + LocalDateTime.now());
-        }
-        var item = itemService.getItem(bookingDtoCreate.getItemId());
-        if (item == null) {
-            throw new NotFoundParametrException("при бронировании, была указана несуществующая вещь");
-        }
-        if (!item.getAvailable()) {
-            throw new BadParametrException("при бронировании запрошена вещь со статусом Available = false. Item = "
-                    + item);
-        }
-        if (userService.getUserById(userId).isEmpty()) {
-            throw new NotFoundParametrException("при бронировании, был указан несуществующий пользователь владелец");
-        }
-        var ownerId = item.getOwner().getId();
-        if (userId.equals(ownerId)) {
-            throw new NotFoundParametrException("пользователь не может забронировать свою вещь");
-        }
-        Booking booking = BookingMapper.toBooking(bookingDtoCreate, item,
-                userService.getUserById(userId).get());
-        return bookingService.saveBooking(booking);
+
+        return bookingService.saveBooking(bookingDtoCreate, userId);
     }
 
     @PatchMapping("/{bookingId}")
     public BookingDto verificationBooking(@PathVariable Long bookingId,
                                           @RequestHeader(value = "X-Sharer-User-Id") @NotNull Long userId,
                                           @RequestParam(value = "approved") Boolean approved) {
-        var booking = bookingService.findBookingById(bookingId);
-        if (booking.isEmpty()) {
-            throw new NotFoundParametrException("отсутствует бронирование с id = " + bookingId);
-        }
-        if (!booking.get().getItem().getOwner().getId().equals(userId)) {
-            throw new NotFoundParametrException("Редактировать статус бронирования может только владелец вещи");
-        }
-        return bookingService.updateApproved(bookingId, approved);
+
+        return bookingService.updateApproved(bookingId, approved, userId);
     }
 
     @GetMapping("/{bookingId}")
     public BookingDto getBooking(@PathVariable Long bookingId,
                                  @RequestHeader(value = "X-Sharer-User-Id") @NotNull Long userId) {
-        var booking = bookingService.findBookingById(bookingId);
-        if (booking.isEmpty()) {
-            throw new NotFoundParametrException("отсутствует бронирование с id = " + bookingId);
-        }
-        if (booking.get().getBooker().getId().equals(userId) || booking.get().getItem().getOwner().getId().equals(userId)) {
-            return BookingMapper.toBookingDto(booking.get());
-        } else {
-            throw new NotFoundParametrException("запрашивать бронирование может или автор или владелец вещи");
-        }
+
+        return bookingService.findBookingById(bookingId, userId);
     }
 
     @GetMapping
     public List<BookingDto> findAllBookingWithStatus(
-            @RequestParam(value = "state", defaultValue = "ALL", required = false) String state,
+            @RequestParam(value = "state", defaultValue = "ALL") String state,
             @RequestHeader(value = "X-Sharer-User-Id") @NotNull Long userId) {
-        if (userService.getUserById(userId).isEmpty()) {
-            throw new NotFoundParametrException("Запрошены бронирования от несуществующего пользователя. Id = " + userId);
-        }
+
         return bookingService.findAllBookingWithStatus(userId, state);
     }
 
     @GetMapping("/owner")
     public List<BookingDto> findAllBookingForUserWithStatus(
-            @RequestParam(value = "state", defaultValue = "ALL", required = false) String state,
+            @RequestParam(value = "state", defaultValue = "ALL") String state,
             @RequestHeader(value = "X-Sharer-User-Id") @NotNull Long userId) {
-        if (userService.getUserById(userId).isEmpty()) {
-            throw new NotFoundParametrException("Запрошены бронирования от несуществующего пользователя. Id = " + userId);
-        }
-        if (bookingService.countItemForUser(userId) == 0L) {
-            return new ArrayList<>();
-        }
+
         return bookingService.findAllBookingForUserWithStatus(userId, state);
     }
