+++ src/main/java/ru/practicum/shareit/user/UserController.java
@@ -1,7 +1,9 @@
 package ru.practicum.shareit.user;
 
+import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.shareit.exceptions.BadParametrException;
@@ -17,36 +19,25 @@ import java.util.List;
 @RequestMapping(path = "/users")
 @Validated
 @Slf4j
+@AllArgsConstructor
 public class UserController {
     private final UserService userService;
 
-    @Autowired
-    UserController(UserService userService) {
-        this.userService = userService;
-    }
-
     @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
     public UserDto createUser(@Valid @RequestBody User user) {
         return userService.createUser(user);
     }
 
     @PatchMapping("/{userId}")
     public UserDto patchUser(@RequestBody @NotNull User user, @PathVariable @NotNull Long userId) {
-        if (userService.getUserById(userId).isEmpty()) {
-            log.info("Попытка запросить редактирование отсутствующего пользователя. userId= {}", userId);
-            throw new BadParametrException("Отсутствует запрашиваемый пользователь. userId= " + userId);
-        }
 
         return userService.patchUser(user, userId);
     }
 
     @GetMapping("/{userId}")
     public UserDto getUserById(@PathVariable Long userId) {
-        var rez = userService.getUserById(userId);
-        if (rez.isEmpty()) {
-            throw new NotFoundParametrException("Отсутствует запрашиваемый пользователь. userId= " + userId);
-        }
-        return UserMapper.toUserDto(rez.get());
+        return userService.getUserById(userId);
     }
 
     @GetMapping
